move utility classes to `dev.architectury.loom.util`
