+++ src/main/java/dev/architectury/loom/util/legacyforge/CoreModManagerTransformer.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package net.fabricmc.loom.util.legacyforge;
+package dev.architectury.loom.util.legacyforge;
 
 import static org.objectweb.asm.Opcodes.ACC_PRIVATE;
 import static org.objectweb.asm.Opcodes.ACC_STATIC;
