+++ src/main/java/net/fabricmc/loom/configuration/providers/forge/fg2/LegacyPatchedProvider.java
@@ -39,6 +39,8 @@ import java.util.Objects;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
+import dev.architectury.loom.util.AccessTransformSetMapper;
+import dev.architectury.loom.util.legacyforge.CoreModManagerTransformer;
 import com.google.common.base.Stopwatch;
 import org.cadixdev.at.AccessTransformSet;
 import org.cadixdev.at.io.AccessTransformFormats;
@@ -52,14 +54,12 @@ import net.fabricmc.loom.configuration.providers.forge.ForgeProvider;
 import net.fabricmc.loom.configuration.providers.forge.MinecraftPatchedProvider;
 import net.fabricmc.loom.configuration.providers.forge.PatchProvider;
 import net.fabricmc.loom.configuration.providers.mappings.MappingConfiguration;
-import net.fabricmc.loom.configuration.providers.mappings.utils.AccessTransformSetMapper;
 import net.fabricmc.loom.configuration.providers.minecraft.MinecraftJarMerger;
 import net.fabricmc.loom.configuration.providers.minecraft.MinecraftProvider;
 import net.fabricmc.loom.util.FileSystemUtil;
 import net.fabricmc.loom.util.Pair;
 import net.fabricmc.loom.util.TinyRemapperHelper;
 import net.fabricmc.loom.util.ZipUtils;
-import net.fabricmc.loom.util.legacyforge.CoreModManagerTransformer;
 import net.fabricmc.lorenztiny.TinyMappingsReader;
 import net.fabricmc.mappingio.MappingReader;
 import net.fabricmc.mappingio.tree.MemoryMappingTree;
